refactor(): move rules into central directory to allow for later entity usage
